Removing the account ids hardcoded in the repo.
This lead to removing the sqs queue listening to the Dead letter SNS topic.

The original intention of having this Queue is to make it a error retry loop. But in reality can't find anything really going into it or we have implemented it.

In the real world usage at the moment, SNS topic is listened by emails and altering systems (notifying to slack etc..)

Other changes are to enable testing.

This is tested by packaging and removing using serverless commands